LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-row-002-dynamic-expected.txt

Visibility collapse

Spec
Setting a row to visibility:collapse changes table height but not width.

row 1	
row 2	
Bottom table is identical to top except row 2 has been collapsed.
row 1	

PASS row visibility:collapse doesn't change table width, unlike display:none
FAIL row visibility:collapse changes table height, unlike visibility:hidden assert_equals: row visibility:collapse changes table height, unlike visibility:hidden expected 116 but got 222
